refactor: enhance reshare vault flow and components
- Introduced `ReshareSecureVaultFlow` and `ReshareVerifyStep` components to streamline the reshare process.
- Updated `SecureMigrateVaultPage` and `SecureReshareVaultPage` to utilize the new reshare components, improving modularity.
- Cleaned up import statements by removing obsolete references and ensuring proper organization.
- Added `InfoBlock` component for better status messaging in various pages.
