+++ clients/desktop/src/vault/keygen/shared/FastVaultKeygenFlow.tsx
@@ -1,6 +1,7 @@
 import { hasServer, isServer } from '@core/mpc/devices/localPartyId'
 import { KeygenFlow } from '@core/ui/mpc/keygen/flow/KeygenFlow'
 import { KeygenPeerDiscoveryStep } from '@core/ui/mpc/keygen/peers/KeygenPeerDiscoveryStep'
+import { ReshareVerifyStep } from '@core/ui/mpc/keygen/reshare/verify/ReshareVerifyStep'
 import { StartMpcSessionStep } from '@core/ui/mpc/session/StartMpcSessionStep'
 import { useCurrentVault } from '@core/ui/vault/state/currentVault'
 import { Match } from '@lib/ui/base/Match'
@@ -9,7 +10,6 @@ import { useNavigateBack } from '@lib/ui/navigation/hooks/useNavigateBack'
 
 import { MpcMediatorManager } from '../../../mpc/serverType/MpcMediatorManager'
 import { JoinKeygenSessionStep } from '../../keygen/shared/JoinKeygenSessionStep'
-import { ReshareVerifyStep } from '../../reshare/shared/ReshareVerifyStep'
 import { ServerEmailStep } from '../../server/email/ServerEmailStep'
 import { ServerPasswordStep } from '../../server/password/ServerPasswordStep'
 import { SetServerPasswordStep } from '../../server/password/SetServerPasswordStep'
