+++ clients/desktop/src/vault/reshare/ReshareVaultPage.tsx
@@ -5,11 +5,11 @@ import { PageContent } from '@lib/ui/page/PageContent'
 import { PageHeader } from '@lib/ui/page/PageHeader'
 import { PageHeaderBackButton } from '@lib/ui/page/PageHeaderBackButton'
 import { PageHeaderTitle } from '@lib/ui/page/PageHeaderTitle'
+import { InfoBlock } from '@lib/ui/status/InfoBlock'
 import { Text } from '@lib/ui/text'
 import { match } from '@lib/utils/match'
 import { useTranslation } from 'react-i18next'
 
-import { InfoBlock } from '../../lib/ui/status/InfoBlock'
 import { useAppNavigate } from '../../navigation/hooks/useAppNavigate'
 
 export const ReshareVaultPage = () => {
