+++ clients/desktop/src/vault/reshare/secure/SecureReshareVaultPage.tsx
@@ -1,15 +1,17 @@
+import { ReshareSecureVaultFlow } from '@core/ui/mpc/keygen/reshare/ReshareSecureVaultFlow'
 import { ReshareVaultFlowProviders } from '@core/ui/mpc/keygen/reshare/ReshareVaultFlowProviders'
 import { CurrentKeygenTypeProvider } from '@core/ui/mpc/keygen/state/currentKeygenType'
 
+import { MpcMediatorManager } from '../../../mpc/serverType/MpcMediatorManager'
 import { ReshareVaultKeygenActionProvider } from '../../keygen/reshare/ReshareVaultKeygenActionProvider'
-import { SecureVaultKeygenFlow } from '../../keygen/shared/SecureVaultKeygenFlow'
 
 export const SecureReshareVaultPage = () => {
   return (
     <ReshareVaultFlowProviders>
       <CurrentKeygenTypeProvider value="reshare">
         <ReshareVaultKeygenActionProvider>
-          <SecureVaultKeygenFlow />
+          <MpcMediatorManager />
+          <ReshareSecureVaultFlow />
         </ReshareVaultKeygenActionProvider>
       </CurrentKeygenTypeProvider>
     </ReshareVaultFlowProviders>
