+++ clients/desktop/src/vault/server/password/ServerPasswordStep.tsx
@@ -11,12 +11,11 @@ import { PageHeader } from '@lib/ui/page/PageHeader'
 import { PageHeaderBackButton } from '@lib/ui/page/PageHeaderBackButton'
 import { PageHeaderTitle } from '@lib/ui/page/PageHeaderTitle'
 import { OnFinishProp } from '@lib/ui/props'
+import { InfoBlock } from '@lib/ui/status/InfoBlock'
 import { Text } from '@lib/ui/text'
 import { useMutation } from '@tanstack/react-query'
 import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
-
-import { InfoBlock } from '../../../lib/ui/status/InfoBlock'
 
 export const ServerPasswordStep: React.FC<OnFinishProp> = ({ onFinish }) => {
   const { t } = useTranslation()
