+++ core/ui/mpc/keygen/reshare/verify/ReshareDeviceItem.tsx
